Show loading progress to help with large folders

+++ utils/utils.go
@@ -25,14 +25,14 @@ func FormatSize(sizeBytes uint64, human bool) string {
 	}
 }
 
-func Max(x, y int64) int64 {
+func Max(x, y int) int {
 	if x >= y {
 		return x
 	}
 	return y
 }
 
-func Min(x, y int64) int64 {
+func Min(x, y int) int {
 	if x <= y {
 		return x
 	}
